Renamed slate_base to trossen_slate, added/modified comments, renamed demos to demo for consistency with other repos, removed period from CMakeLists.txt.
